Выносит компоненты из ForumTopic
