+++ packages/client/src/pages/ForumTopic/components/styles.module.scss
@@ -1,0 +1,89 @@
+@use 'themes/mixins' as mx;
+
+.comments {
+    &-container {
+        display: flex;
+        width: 100%;
+        height: 45rem;
+    }
+
+    &__list {
+        display: block;
+        overflow-y: scroll;
+    }
+
+    &-item {
+        display: flex;
+        min-height: 12rem;
+
+        &-icon {
+            width: 7rem;
+            height: 7rem;
+        }
+
+        &-text {
+            flex-basis: fit-content;
+            flex-grow: 1;
+            margin: 2rem;
+
+            &>span {
+                display: flex;
+                font-size: 1.5rem;
+                margin-bottom: 1rem;
+            }
+
+            &>p {
+                @include mx.themify('color', 'p-color');
+                font-size: 1.2rem;
+            }
+        }
+
+        &-date {
+            @include mx.themify('color', 'forum_date-color');
+            display: flex;
+            height: 10rem;
+            text-align: end;
+        }
+    }
+}
+
+.topicpage__form {
+    display: flex;
+    flex-direction: column;
+
+    &-textfield {
+        @include mx.themify('background-color', 'forum_topic-bg');
+
+        textarea {
+            font-size: 1.4rem;
+        }
+    }
+
+    .buttons-container {
+        margin-top: 2rem;
+        justify-content: end;
+        display: flex;
+
+
+        svg {
+            @include mx.themify('fill', 'forum_backlink-color');
+        }
+    }
+
+    &>button {
+        width: 20rem;
+        margin-top: 1rem;
+        align-self: flex-end;
+    }
+}
+
+.emoji-picker {
+    position: absolute;
+    top: 32rem;
+    left: 120rem;
+
+    &-container {
+        position: relative;
+        z-index: 9999;
+    }
+}
